:truck: [#435] -- group CreateAppointment component code
