+++ src/components/appointments/CreateAppointment/CreateAppointment.stories.js
@@ -7,14 +7,14 @@ import {buildForm} from 'api-mocks';
 import {mockSubmissionPost} from 'api-mocks/submissions';
 import {ConfigDecorator, LayoutDecorator} from 'story-utils/decorators';
 
-import CreateAppointment, {routes as childRoutes} from './CreateAppointment';
 import {
   mockAppointmentCustomerFieldsGet,
   mockAppointmentDatesGet,
   mockAppointmentLocationsGet,
   mockAppointmentProductsGet,
   mockAppointmentTimesGet,
-} from './mocks';
+} from '../mocks';
+import CreateAppointment, {routes as childRoutes} from './';
 
 export default {
   title: 'Private API / Appointments / CreateForm',
