+++ src/components/appointments/CreateAppointment/index.js
@@ -1,0 +1,4 @@
+import CreateAppointment from './CreateAppointment';
+
+export default CreateAppointment;
+export {routes} from './routes';
